Extract queue rows to stay under the module-size cap
JobRow, the published shelf, and confirm copy live in their own modules so
AdminJobsPanel can shrink. Confirm coverage moves to a dedicated test file.
