+++ apps/web/src/AdminJobsPanel.test.tsx
@@ -55,16 +55,6 @@ async function render() {
   return { container, root };
 }
 
-function confirmDialog(): HTMLElement | null {
-  return document.querySelector('.admin-job-confirm');
-}
-
-function confirmDialogButton(label: string): HTMLButtonElement {
-  return Array.from(confirmDialog()?.querySelectorAll('button') ?? []).find(
-    (button) => button.textContent === label,
-  ) as HTMLButtonElement;
-}
-
 afterEach(() => {
   document.body.innerHTML = '';
   vi.clearAllMocks();
@@ -80,100 +70,6 @@ describe('AdminJobsPanel', () => {
     const { container, root } = await render();
 
     expect(container.querySelectorAll('.admin-job-publish')).toHaveLength(1);
-
-    await act(async () => root.unmount());
-  });
-
-  it('publishes and says which game went live', async () => {
-    mocked.fetchJobQueue.mockResolvedValue(queue([job()]));
-    mocked.publishJob.mockResolvedValue({
-      ok: true,
-      slug: 'comet-courier',
-      version: 'v1',
-      publishedAt: '2026-07-30T12:00:00Z',
-    });
-
-    const { container, root } = await render();
-    const button = container.querySelector('.admin-job-publish') as HTMLButtonElement;
-
-    await act(async () => {
-      button.click();
-      await Promise.resolve();
-    });
-
-    expect(mocked.publishJob).not.toHaveBeenCalled();
-    expect(confirmDialog()?.textContent).toContain('Publish Comet Courier?');
-    expect(confirmDialog()?.textContent).toContain('comet-courier');
-
-    await act(async () => {
-      confirmDialogButton('Publish').click();
-      await Promise.resolve();
-    });
-
-    expect(mocked.publishJob).toHaveBeenCalledWith(1_000_001);
-    expect(container.querySelector('.admin-job-message')?.textContent).toContain('published comet-courier');
-    expect(mocked.fetchJobQueue).toHaveBeenCalledTimes(2);
-
-    await act(async () => root.unmount());
-  });
-
-  it('does not publish when the confirmation is dismissed', async () => {
-    mocked.fetchJobQueue.mockResolvedValue(queue([job()]));
-
-    const { container, root } = await render();
-    await act(async () => {
-      (container.querySelector('.admin-job-publish') as HTMLButtonElement).click();
-      await Promise.resolve();
-    });
-
-    await act(async () => {
-      confirmDialogButton('Back').click();
-      await Promise.resolve();
-    });
-
-    expect(mocked.publishJob).not.toHaveBeenCalled();
-    expect(confirmDialog()).toBeNull();
-
-    await act(async () => root.unmount());
-  });
-
-  it('closes the confirmation on Escape without publishing', async () => {
-    mocked.fetchJobQueue.mockResolvedValue(queue([job()]));
-
-    const { container, root } = await render();
-    await act(async () => {
-      (container.querySelector('.admin-job-publish') as HTMLButtonElement).click();
-      await Promise.resolve();
-    });
-    expect(confirmDialog()).not.toBeNull();
-
-    await act(async () => {
-      window.dispatchEvent(new KeyboardEvent('keydown', { key: 'Escape', bubbles: true }));
-      await Promise.resolve();
-    });
-
-    expect(mocked.publishJob).not.toHaveBeenCalled();
-    expect(confirmDialog()).toBeNull();
-
-    await act(async () => root.unmount());
-  });
-
-  it('names the next step when a publish is refused', async () => {
-    mocked.fetchJobQueue.mockResolvedValue(queue([job()]));
-    mocked.publishJob.mockResolvedValue({ refused: 'gate_red' });
-
-    const { container, root } = await render();
-
-    await act(async () => {
-      (container.querySelector('.admin-job-publish') as HTMLButtonElement).click();
-      await Promise.resolve();
-    });
-    await act(async () => {
-      confirmDialogButton('Publish').click();
-      await Promise.resolve();
-    });
-
-    expect(container.querySelector('.admin-job-message')?.textContent).toContain('gate failed');
 
     await act(async () => root.unmount());
   });
@@ -217,67 +113,6 @@ describe('AdminJobsPanel', () => {
 
     expect(container.querySelector('.admin-job-stall')?.textContent).toContain('silent');
     expect(container.querySelector('.admin-job-row')?.className).toContain('is-stalled');
-
-    await act(async () => root.unmount());
-  });
-
-  it('takes a confirmation dialog to cancel, because canceled has no undo', async () => {
-    mocked.fetchJobQueue.mockResolvedValue(queue([job({ state: 'building' })]));
-    mocked.cancelJob.mockResolvedValue({ ok: true, state: 'canceled', stopEnforced: false });
-
-    const { container, root } = await render();
-
-    // Switch to in-flight filter to see building job
-    const inFlightChip = Array.from(container.querySelectorAll('.admin-filter-chip')).find((c) =>
-      c.textContent?.includes('In flight'),
-    ) as HTMLButtonElement;
-    await act(async () => {
-      inFlightChip.click();
-      await Promise.resolve();
-    });
-
-    const button = container.querySelector('.admin-job-cancel') as HTMLButtonElement;
-    await act(async () => {
-      button.click();
-      await Promise.resolve();
-    });
-    expect(mocked.cancelJob).not.toHaveBeenCalled();
-    expect(confirmDialog()?.textContent).toContain('Cancel Comet Courier?');
-    expect(confirmDialog()?.textContent).toContain('cannot be undone');
-
-    await act(async () => {
-      confirmDialogButton('Cancel build').click();
-      await Promise.resolve();
-    });
-    expect(mocked.cancelJob).toHaveBeenCalledWith(1_000_001);
-    expect(container.querySelector('.admin-job-message')?.textContent).toContain('next report');
-
-    await act(async () => root.unmount());
-  });
-
-  it('does not cancel when the confirmation is dismissed', async () => {
-    mocked.fetchJobQueue.mockResolvedValue(queue([job({ state: 'building' })]));
-
-    const { container, root } = await render();
-    const inFlightChip = Array.from(container.querySelectorAll('.admin-filter-chip')).find((c) =>
-      c.textContent?.includes('In flight'),
-    ) as HTMLButtonElement;
-    await act(async () => {
-      inFlightChip.click();
-      await Promise.resolve();
-    });
-
-    await act(async () => {
-      (container.querySelector('.admin-job-cancel') as HTMLButtonElement).click();
-      await Promise.resolve();
-    });
-    await act(async () => {
-      confirmDialogButton('Keep').click();
-      await Promise.resolve();
-    });
-
-    expect(mocked.cancelJob).not.toHaveBeenCalled();
-    expect(confirmDialog()).toBeNull();
 
     await act(async () => root.unmount());
   });
@@ -554,86 +389,6 @@ describe('AdminJobsPanel', () => {
     });
 
     expect(document.querySelector('.admin-preview-modal')).toBeNull();
-
-    await act(async () => root.unmount());
-  });
-
-  it('confirms publish from the preview theater before going live', async () => {
-    mocked.fetchJobQueue.mockResolvedValue(queue([job({ jobId: 1234, title: 'Sky Dodge', slug: 'sky-dodge' })]));
-    mocked.fetchJobPreview.mockResolvedValue({
-      slug: 'sky-dodge',
-      title: 'Sky Dodge',
-      version: 'v1',
-      html: '<!doctype html><html><body>Playable Game Canvas</body></html>',
-    });
-    mocked.publishJob.mockResolvedValue({
-      ok: true,
-      slug: 'sky-dodge',
-      version: 'v1',
-      publishedAt: '2026-07-30T12:00:00Z',
-    });
-
-    const { container, root } = await render();
-    await act(async () => {
-      (container.querySelector('.admin-job-preview-btn') as HTMLButtonElement).click();
-      await Promise.resolve();
-      await Promise.resolve();
-    });
-
-    const publishBtn = Array.from(document.querySelectorAll('.admin-preview-actions button')).find(
-      (button) => button.textContent === 'Publish Game',
-    ) as HTMLButtonElement;
-    await act(async () => {
-      publishBtn.click();
-      await Promise.resolve();
-    });
-
-    expect(mocked.publishJob).not.toHaveBeenCalled();
-    expect(confirmDialog()?.textContent).toContain('Publish Sky Dodge?');
-    expect(document.querySelector('.admin-preview-modal')).not.toBeNull();
-
-    await act(async () => {
-      confirmDialogButton('Publish').click();
-      await Promise.resolve();
-    });
-    expect(mocked.publishJob).toHaveBeenCalledWith(1234);
-
-    await act(async () => root.unmount());
-  });
-
-  it('executes batch publish on deduplicated ready jobs', async () => {
-    mocked.fetchJobQueue.mockResolvedValue(
-      queue([
-        job({ jobId: 10, slug: 'game-1', state: 'ready_for_review' }),
-        job({ jobId: 9, slug: 'game-1', state: 'ready_for_review' }), // older superseded build
-        job({ jobId: 11, slug: 'game-2', state: 'ready_for_review' }),
-        job({ jobId: 12, slug: 'game-3', state: 'building' }),
-      ]),
-    );
-    mocked.publishJob.mockResolvedValue({ ok: true, slug: 'game', version: 'v1', publishedAt: '2026-07-30T12:00:00Z' });
-
-    const { container, root } = await render();
-    const bulkBtn = container.querySelector('.admin-bulk-publish-cta') as HTMLButtonElement;
-    expect(bulkBtn?.textContent).toContain('Publish all ready (2)');
-
-    await act(async () => {
-      bulkBtn.click();
-      await Promise.resolve();
-    });
-
-    expect(mocked.publishJob).not.toHaveBeenCalled();
-    expect(confirmDialog()?.textContent).toContain('Publish 2 games?');
-
-    await act(async () => {
-      confirmDialogButton('Publish 2').click();
-      await Promise.resolve();
-      await Promise.resolve();
-    });
-
-    expect(mocked.publishJob).toHaveBeenCalledWith(10);
-    expect(mocked.publishJob).toHaveBeenCalledWith(11);
-    expect(mocked.publishJob).not.toHaveBeenCalledWith(9);
-    expect(mocked.publishJob).not.toHaveBeenCalledWith(12);
 
     await act(async () => root.unmount());
   });
